Various small value changes and changed the world box from cube to 4 planes
Changed how far back the camera moves. Changed the world box as it was not showing up on my phone when built as APK. Changed player movement to be faster as user feedback has been that it was too diffiuclt to move across the lanes quickly. Changed fog to be less dense but on my phone it is still too foggy. Will need to discuss